% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_count_value.R
\name{compute_count_value}
\alias{compute_count_value}
\title{Compute count value}
\usage{
compute_count_value(play, event, base_out_run_exp)
}
\arguments{
\item{play}{dataframe of play data from \code{\link{download_statsapi}}}

\item{event}{dataframe of event data from \code{\link{download_statsapi}}}

\item{base_out_run_exp}{dataframe of base-out run expectancy from
\code{\link{compute_base_out_run_exp}}}
}
\value{
a dataframe of \code{count_value} indexed by \code{balls} and \code{strikes}
}
\description{
Compute the average change in base-out run expectancy for all plate appearances that run
through each count.
}
