% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_40man_roster.R
\name{download_40man_roster}
\alias{download_40man_roster}
\title{Download MLB 40-Man Rosters}
\usage{
download_40man_roster(date = Sys.Date(), cl = NULL)
}
\arguments{
\item{date}{A character string representing the date in \code{YYYY-MM-DD} format.
Defaults to the current system date.}

\item{cl}{optional cluster object for parallel computation, default is NULL (not parallel)}
}
\value{
A \code{tibble} containing the following columns:
\itemize{
\item \strong{team_id}: The unique numeric identifier for the MLB team.
\item \strong{player_id}: The unique numeric identifier for the player.
\item \strong{status}: The character code representing the player's current roster status.
Possible values include:
\itemize{
\item A: Active
\item BRV: Bereavement List
\item D{#}: Injured {#}-Day
\item PL: Paternity List
\item RM: Reassigned to Minors
\item SU: Suspended
}
}
}
\description{
Retrieves the 40-man roster for all Major League Baseball teams for a specified date
using the MLB Stats API.
}
