% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_player.R
\name{download_player}
\alias{download_player}
\title{Download player table}
\usage{
download_player(
  year,
  level = c("MLB", "AAA", "AA", "A+", "A", "CL", "DSL"),
  cl = NULL
)
}
\arguments{
\item{year}{vector of integers (first valid year is 1876)}

\item{level}{character vector of levels. Valid levels are:
\itemize{
\item{"MLB" for Major League Baseball}
\item{"AAA" for Triple-A}
\item{"AA" for Double-A}
\item{"A+" for High-A}
\item{"A" for Single-A}
\item{"SS" (defunct since 2020) for Short-Season A and Rookie Advanced}
\item{"CL" for Complex Leagues}
\item{"DSL" for Dominican Summer League}
}}

\item{cl}{optional cluster object for parallel computation, default is NULL (not parallel)}
}
\value{
a table of player data, indexed by player_id
}
\description{
Download player information for a vector of years and a vector of levels.
}
