% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_schedule.R
\name{download_schedule}
\alias{download_schedule}
\title{Download schedule}
\usage{
download_schedule(
  start_date,
  end_date,
  level = c("MLB", "AAA", "AA", "A+", "A"),
  game_type = "R"
)
}
\arguments{
\item{start_date}{first date included in the download}

\item{end_date}{last date included in the download}

\item{level}{character vector of levels. Valid levels are:
\itemize{
\item{"MLB" for Major League Baseball}
\item{"AAA" for Triple-A}
\item{"AA" for Double-A}
\item{"A+" for High-A}
\item{"A" for Single-A}
\item{"SS" (defunct since 2020) for Short-Season A and Rookie Advanced}
\item{"CL" for Complex Leagues}
\item{"DSL" for Dominican Summer League}
}}

\item{game_type}{character vector of game types. Valid game types are:
\itemize{
\item{"R" for Regular Season}
\item{"F" for First-Round Playoff Series (aka wild card)}
\item{"D" for Division Series (or minor-league quarter-finals)}
\item{"L" for League Championship Series (or minor league semi-finals)}
\item{"W" for World Series (or minor league championship)}
\item{"A" for All-Star Game}
\item{"E" for Exhibition}
}}
}
\value{
a dataframe of games, with columns \code{game_id}, \code{year}, \code{date}, \code{team_id_away}, \code{team_id_home}, \code{venue_id}
}
\description{
Download a table of MLB or AAA games within a specified date range (must be same calendar year).
}
