% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_season_summary.R
\name{download_season_summary}
\alias{download_season_summary}
\title{Download player stats season summary}
\usage{
download_season_summary(
  year,
  level = "MLB",
  position = c("hitting", "pitching"),
  game_type = c("R", "P"),
  cl = NULL
)
}
\arguments{
\item{year}{integer vector of years}

\item{level}{character vector of levels. Valid levels are:
\itemize{
\item{"MLB" for Major League Baseball}
\item{"AAA" for Triple-A}
\item{"AA" for Double-A}
\item{"A+" for High-A}
\item{"A" for Single-A}
\item{"SS" (defunct since 2020) for Short-Season A and Rookie Advanced}
\item{"CL" for Complex Leagues}
\item{"DSL" for Dominican Summer League}
}}

\item{position}{"hitting" or "pitching" (cannot be both)}

\item{game_type}{"R" for regular season or "P" for playoffs (cannot be both)}

\item{cl}{optional cluster object for parallel computation, default is NULL (not parallel)}
}
\value{
a table of season summary statistics indexed by year, level, league_id and player_id
}
\description{
This function this the https://bdfed.stitch.mlbinfra.com/bdfed/stats/player api to download
player season summary stats. Coverage goes back to 1901 for MLB, 1990 for domestic minor leagues
and 2016 for the Dominican Summer League. Year and level can be provided as vector arugments.
}
