% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_base_out_run_exp.R
\name{estimate_base_out_run_exp}
\alias{estimate_base_out_run_exp}
\title{Estimate base-out run expectancy}
\usage{
estimate_base_out_run_exp(event)
}
\arguments{
\item{event}{dataframe of event data from \code{\link{download_statsapi}}}
}
\value{
a dataframe of \code{exp_runs} indexed by \code{runner_1b}, \code{runner_2b}, \code{runner_3b} and \code{out}
}
\description{
Model base-out transitions as a Markov chain and find the probability of each terminal state
(including runs scored) from any initial state.
}
