% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fielding_lineup.R
\name{extract_fielding_lineup}
\alias{extract_fielding_lineup}
\title{Extract fielding lineup}
\usage{
extract_fielding_lineup(players)
}
\arguments{
\item{players}{a players JSON object extracted from a boxscore JSON object}
}
\value{
a table of player_id keyed on position (2-10)
}
\description{
Extract the starting fielding lineup from a provided players JSON object
}
