% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_game.R
\name{extract_game}
\alias{extract_game}
\title{Extract game}
\usage{
extract_game(game_id)
}
\arguments{
\item{game_id}{an integer primary key specifying the game to extract (character is okay)}
}
\value{
a list of two dataframes: \code{event} and \code{pitch}
}
\description{
Extract event and pitch data from the MLB statsapi.
}
