% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_level_by_league_year.R
\name{get_league_level_by_year}
\alias{get_league_level_by_year}
\title{Get league level by year}
\usage{
get_league_level_by_year(year)
}
\arguments{
\item{year}{integer vector of years}
}
\value{
a table of \code{level} indexed by \code{league_id} and \code{year}
}
\description{
The minor leagues have gone through significant restructuring a couple of times (1990 and 2020),
so we can't statically map each league to a level. This function takes a vector of years and
returns a mapping from league_id to level for each year.
}
