% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quadratic_coef.R
\name{get_quadratic_coef}
\alias{get_quadratic_coef}
\title{Get quadratic coefficients for pitch data}
\usage{
get_quadratic_coef(data, source = c("statsapi", "baseballsavant"))
}
\arguments{
\item{data}{a dataframe with the columns required for calculating the quadratic coefficients.
If \code{source = "statsapi"}, then \code{ax}, \code{ay}, \code{az}, \code{vx0}, \code{vy0}, \code{vz0}, \code{x0}, \code{z0}, and
\code{extension} are required. If \code{source = "baseballsavant"}, then \code{ax}, \code{ay}, \code{az}, \code{vx0}, \code{vy0},
\code{vz0}, \code{release_pos_x}, \code{release_pos_y}, and \code{release_pos_z} are required.}

\item{source}{character string, either "statsapi" (default) or "baseballsavant"}
}
\value{
the input dataframe the following columns added: bx, by, bz, cx, cy, cz, y0, t0
}
\description{
Find coefficients such that the path of the ball through time is given by:
x(t) = ax * t^2 / 2 + bx * t + cx
y(t) = ay * t^2 / 2 + by * t + cy
z(t) = az * t^2 / 2 + bz * t + cz
We start with velocity and location of ball when y = 50 and use extension to derive coefficients
}
