% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sport_id.R
\name{get_sport_id}
\alias{get_sport_id}
\title{Get Sport ID for a specified level}
\usage{
get_sport_id(level)
}
\arguments{
\item{level}{character vector of levels. Valid levels are:
\itemize{
\item{"MLB" for Major League Baseball}
\item{"AAA" for Triple-A}
\item{"AA" for Double-A}
\item{"A+" for High-A}
\item{"A" for Single-A}
\item{"SS" (defunct since 2020) for Short-Season A and Rookie Advanced}
\item{"CL" for Complex Leagues}
\item{"DSL" for Dominican Summer League}
}}
}
\value{
an integer vector of sport IDs
}
\description{
This function returned a vector of sport IDs corresponding to a vector of levels.
}
