% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_video_url.R
\name{get_video_url}
\alias{get_video_url}
\title{Get the URL to watch the video of any play}
\usage{
get_video_url(play_id)
}
\arguments{
\item{play_id}{character vector of play identifier from statsapi}
}
\value{
character vector of URLs
}
\description{
Using \code{play_id} from the statsapi download, you can pull up the video of any play
}
\examples{
get_video_url("064e8918-e2b7-40f0-ac39-968932652154")

}
