% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_null.R
\name{replace_null}
\alias{replace_null}
\title{Replace NULL value}
\usage{
replace_null(x, replacement = NA)
}
\arguments{
\item{x}{the value to replace if NULL}

\item{replacement}{the value with which to replace NULL, defaults to NA}
}
\value{
If x is NULL, return replacement, else return x
}
\description{
Replace NULL value with a specified replacement value, useful when an API endpoint may not
return a value for a given field.
}
