% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_game_type}
\alias{sanitize_game_type}
\title{Sanitize game type}
\usage{
sanitize_game_type(game_type)
}
\arguments{
\item{game_type}{character vector of game types. Valid game types are:
\itemize{
\item{"R" for Regular Season}
\item{"F" for First-Round Playoff Series (aka wild card)}
\item{"D" for Division Series (or minor-league quarter-finals)}
\item{"L" for League Championship Series (or minor league semi-finals)}
\item{"W" for World Series (or minor league championship)}
\item{"A" for All-Star Game}
\item{"E" for Exhibition}
}}
}
\value{
simply returns the input vector \code{game_type} but throws error if invalid types are present
}
\description{
Check that a character vector of game types are all valid game types.
}
