% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize_level}
\alias{sanitize_level}
\title{Sanitize level}
\usage{
sanitize_level(level)
}
\arguments{
\item{level}{character vector of levels. Valid levels are:
\itemize{
\item{"MLB" for Major League Baseball}
\item{"AAA" for Triple-A}
\item{"AA" for Double-A}
\item{"A+" for High-A}
\item{"A" for Single-A}
\item{"SS" (defunct since 2020) for Short-Season A and Rookie Advanced}
\item{"CL" for Complex Leagues}
\item{"DSL" for Dominican Summer League}
}}
}
\value{
simply returns the input vector \code{level} but throws error if invalid levels are present
}
\description{
Check that a character vector of levels are all valid levels.
}
