% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_base_out_by_event.R
\name{track_base_out_by_event}
\alias{track_base_out_by_event}
\title{Track base-out state by event}
\usage{
track_base_out_by_event(event_data)
}
\arguments{
\item{event_data}{This is the liveData$plays$allPlays section of the JSON from a game's GUMBO feed}
}
\value{
a tibble with the following columns:
event_index,
pre_runner_1b_id, pre_runner_2b_id, pre_runner_3b_id, pre_out,
post_runner_1b_id, post_runner_2b_id, post_runner_3b_id, post_out
}
\description{
Post-event base-out state is given to us for free. We get pre-event base-out state based on the
post-event base-out state of the previous event.
}
