% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_baseballsavant.R
\name{download_baseballsavant}
\alias{download_baseballsavant}
\title{Download MLB data from baseballsavant.mlb.com}
\usage{
download_baseballsavant(start_date, end_date, cl = NULL, verbose = TRUE)
}
\arguments{
\item{start_date}{first date included in the download}

\item{end_date}{last date included in the download}

\item{cl}{optional cluster object for parallel computation, default is NULL (not parallel)}

\item{verbose}{logical, should progress be printed to console?}
}
\value{
a dataframe with one row per pitch and all available columns
}
\description{
Loop over five days at a time to download data from the Statcast search API at Baseball Savant.
This includes swing tracking data not available through the MLB statsapi.
}
\examples{
\dontrun{
   data_statsapi <- download_statsapi(
     start_date = "2024-07-01",
     end_date = "2024-07-01"
   )
}

}
