% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quadratic_coef.R
\name{get_quadratic_coef}
\alias{get_quadratic_coef}
\title{Get quadratic coefficients for pitch data}
\usage{
get_quadratic_coef(data)
}
\arguments{
\item{data}{a dataframe with the following columns: ax, ay, az, vx0, vy0, vz0, x0, z0, extension}
}
\value{
the input dataframe the following columns added: cy, y0, t0, bx, by, bz, cx, cz
}
\description{
Find coefficients such that the path of the ball through time is given by:
x(t) = ax * t^2 / 2 + bx * t + cx
y(t) = ay * t^2 / 2 + by * t + cy
z(t) = az * t^2 / 2 + bz * t + cz
We start with velocity and location of ball when y = 50 and use extension to derive coefficients
}
