% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_game_type.R
\name{sanitize_game_type}
\alias{sanitize_game_type}
\title{Sanitize game type}
\usage{
sanitize_game_type(game_type)
}
\arguments{
\item{game_type}{character vector of game types}
}
\value{
simply returns the input vector \code{game_type} but throws error if invalid types are present
}
\description{
Check that a character vector of game types are all valid game types.
}
